+++ sdk/resourcemanager/securityinsights/armsecurityinsights/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armsecurityinsights
 
 const (
 	moduleName    = "armsecurityinsights"
-	moduleVersion = "v0.3.0"
+	moduleVersion = "v0.4.0"
 )
 
 // ActionType - The type of the automation rule action
@@ -1007,6 +1007,8 @@ type EntityTimelineKind string
 const (
 	// EntityTimelineKindActivity - activity
 	EntityTimelineKindActivity EntityTimelineKind = "Activity"
+	// EntityTimelineKindAnomaly - anomaly
+	EntityTimelineKindAnomaly EntityTimelineKind = "Anomaly"
 	// EntityTimelineKindBookmark - bookmarks
 	EntityTimelineKindBookmark EntityTimelineKind = "Bookmark"
 	// EntityTimelineKindSecurityAlert - security alerts
@@ -1017,6 +1019,7 @@ const (
 func PossibleEntityTimelineKindValues() []EntityTimelineKind {
 	return []EntityTimelineKind{
 		EntityTimelineKindActivity,
+		EntityTimelineKindAnomaly,
 		EntityTimelineKindBookmark,
 		EntityTimelineKindSecurityAlert,
 	}
@@ -1168,17 +1171,17 @@ func PossibleFileHashAlgorithmValues() []FileHashAlgorithm {
 	}
 }
 
-// GetInsightsErrorKind - the query kind
-type GetInsightsErrorKind string
+// GetInsightsError - the query kind
+type GetInsightsError string
 
 const (
-	GetInsightsErrorKindInsight GetInsightsErrorKind = "Insight"
+	GetInsightsErrorInsight GetInsightsError = "Insight"
 )
 
-// PossibleGetInsightsErrorKindValues returns the possible values for the GetInsightsErrorKind const type.
-func PossibleGetInsightsErrorKindValues() []GetInsightsErrorKind {
-	return []GetInsightsErrorKind{
-		GetInsightsErrorKindInsight,
+// PossibleGetInsightsErrorValues returns the possible values for the GetInsightsError const type.
+func PossibleGetInsightsErrorValues() []GetInsightsError {
+	return []GetInsightsError{
+		GetInsightsErrorInsight,
 	}
 }
 
@@ -1734,6 +1737,20 @@ func PossibleSKUKindValues() []SKUKind {
 	}
 }
 
+// SecurityMLAnalyticsSettingsKind - The kind of security ML analytics settings
+type SecurityMLAnalyticsSettingsKind string
+
+const (
+	SecurityMLAnalyticsSettingsKindAnomaly SecurityMLAnalyticsSettingsKind = "Anomaly"
+)
+
+// PossibleSecurityMLAnalyticsSettingsKindValues returns the possible values for the SecurityMLAnalyticsSettingsKind const type.
+func PossibleSecurityMLAnalyticsSettingsKindValues() []SecurityMLAnalyticsSettingsKind {
+	return []SecurityMLAnalyticsSettingsKind{
+		SecurityMLAnalyticsSettingsKindAnomaly,
+	}
+}
+
 // SettingKind - The kind of the setting
 type SettingKind string
 
@@ -1769,6 +1786,24 @@ func PossibleSettingTypeValues() []SettingType {
 		SettingTypeCopyableLabel,
 		SettingTypeInfoMessage,
 		SettingTypeInstructionStepsGroup,
+	}
+}
+
+// SettingsStatus - The anomaly SecurityMLAnalyticsSettings status
+type SettingsStatus string
+
+const (
+	// SettingsStatusFlighting - Anomaly settings status in Flighting mode
+	SettingsStatusFlighting SettingsStatus = "Flighting"
+	// SettingsStatusProduction - Anomaly settings status in Production mode
+	SettingsStatusProduction SettingsStatus = "Production"
+)
+
+// PossibleSettingsStatusValues returns the possible values for the SettingsStatus const type.
+func PossibleSettingsStatusValues() []SettingsStatus {
+	return []SettingsStatus{
+		SettingsStatusFlighting,
+		SettingsStatusProduction,
 	}
 }